requirements.txt
#Server
Django>=1.5.1
gunicorn
psycopg2==2.4.5
dj-database-url==0.2.0
wsgiref==0.1.2
gevent
South==0.7.4

#Site Specific
Pillow==1.7.7
django-tastypie==0.9.15
django_compressor==1.2
cssmin==0.1.4
jsmin==2.0.2
-e git://github.com/hcarvalhoalves/django-taggit.git@e0f9642d7b94c8e6c0feb520d96bb6ae4d78a4d0#egg=django-taggit

#Cache
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4
django-redis-cache==0.9.2

#Storage
boto==2.2.1
django-queued-storage==0.5
django-storages==1.1.4

#Misc
Markdown==2.0.3
django-picklefield
celery>=2.3.3
django-celery==3.0.10
